requirements.txt
aiobotocore==2.4.1
aiohttp==3.8.3
aioitertools==0.11.0
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.0
backcall==0.2.0
bokeh==2.4.3
botocore==1.27.59
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
comm==0.1.2
dask==2022.12.0
debugpy==1.6.4
decorator==5.1.1
distributed==2022.12.0
entrypoints==0.4
executing==1.2.0
frozenlist==1.3.3
fsspec==2022.11.0
HeapDict==1.0.1
idna==3.4
ipykernel==6.19.2
ipython==8.7.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jupyter_client==7.4.8
jupyter_core==5.1.0
kafka-python==2.0.2
locket==1.0.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
msgpack==1.0.4
multidict==6.0.3
nest-asyncio==1.5.6
numpy==1.23.5
packaging==22.0
pandas==1.5.2
parso==0.8.3
partd==1.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pycodestyle==2.10.0
Pygments==2.13.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
s3fs==2022.11.0
six==1.16.0
sortedcontainers==2.4.0
stack-data==0.6.2
tblib==1.7.0
tomli==2.0.1
toolz==0.12.0
tornado==6.2
traitlets==5.7.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.8.2
zict==2.2.0
